pyproject.toml
[project]
name = "torchreid"
version = "1.4.0"
description = "A library for deep learning person re-ID in PyTorch"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Kaiyang Zhou"}]
keywords = ["Person Re-Identification", "Deep Learning", "Computer Vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21,<2.0",
    "cython>=0.29",
    "h5py>=3.0",
    "pillow>=9.0",
    "scipy>=1.7",
    "opencv-python>=4.5",
    "matplotlib>=3.5",
    "tensorboard>=2.10",
    "yacs>=0.1.8",
    "gdown>=4.0",
    "imageio>=2.9",
    "chardet>=4.0",
    "torch>=2.0",
    "torchvision>=0.15",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
export = [
    "onnx>=1.12",
    "onnx-simplifier>=0.4",
    # Note: openvino-dev may not be available for all Python versions
    # Install manually if needed: pip install openvino-dev
]

[project.urls]
Homepage = "https://github.com/KaiyangZhou/deep-person-reid"
Documentation = "https://kaiyangzhou.github.io/deep-person-reid/"
Repository = "https://github.com/KaiyangZhou/deep-person-reid"

[build-system]
requires = ["setuptools>=61", "wheel", "cython>=0.29", "numpy>=1.21"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["torchreid*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line length handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C408",  # unnecessary dict call
]

[tool.ruff.lint.isort]
known-first-party = ["torchreid"]
force-single-line = false
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
